Corrected calculation of soft limit warning opacity and bug that prevented soft limit warning after hard limit animation.

+++ RsiAlarm/SoftLimitEventArgs.cs
@@ -7,9 +7,9 @@ namespace RsiAlarm
 {
     public class SoftLimitEventArgs : EventArgs
     {
-        public int Level;
+        public double Level;
 
-        public SoftLimitEventArgs(int level)
+        public SoftLimitEventArgs(double level)
         {
             this.Level = level;
         }
